pytest: use pytest_runtest_makereport for consistent exception handling
The `report` passed to `pytest_runtest_logreport` has a different
internal exception representation depending on the `--tb` option with
which `pytest` is configured: some of these representations do not
include the traceback frames to allow us to calculate line numbers.

`pytest_runtest_makereport`, however, has access to the original
`ExceptionInfo` object when an exception is raised: this commit
implements a `pytest_runtest_makereport` hook which simply stores that
object, and modifies `pytest_runtest_logreport` to use it.  This makes
exception handling independent of `--tb` setting.

Fixes: #28